# docs/swagger.yaml
definitions:
...
  handler.authenticateUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handler.authenticateUserResponse:
    properties:
      accessToken:
        type: string
      expiresAt:
        type: string
    type: object
...
  handler.createUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
  handler.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  handler.updateUserRequest:
    properties:
      name:
        type: string
      passwordHash:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  title: BMSTU Web Backend
  version: "1.0"
paths:
...
  /users:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя с указанными именем, электронной почтой
        и паролем.
      parameters:
      - description: Данные для создания пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь успешно создан
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Создание нового пользователя
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Обновляет имя и/или пароль текущего пользователя.
      parameters:
      - description: Данные для обновления пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.updateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Данные пользователя успешно обновлены
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - Bearer: []
      summary: Обновление данных текущего пользователя
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: Аутентифицирует пользователя по электронной почте и паролю и возвращает
        токен доступа.
      parameters:
      - description: Данные для аутентификации пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.authenticateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная аутентификация, токен доступа
          schema:
            $ref: '#/definitions/handler.authenticateUserResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Аутентификация пользователя
      tags:
      - users
  /users/logout:
    post:
      consumes:
      - application/json
      description: Завершает сеанс пользователя, аннулируя токен доступа.
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь успешно вышел
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - Bearer: []
      summary: Выход пользователя
      tags:
      - users
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"